refactor: :ambulance: macro module reorganization and example fixes
